+++ tzbtc.nix
@@ -2,7 +2,8 @@
 #
 # SPDX-License-Identifier: LicenseRef-MIT-BitcoinSuisse
 {
-  pkgs, weeder-hacks, static ? true,
+  pkgs, collect-hie, projectSrc,
+  static ? true,
   release ? false,   # release build with optimizations enabled
   commitInfo ? null  # git commit sha and date
 }:
@@ -12,11 +13,6 @@ with rec {
     if static
     then pkgs.pkgsCross.musl64.haskell-nix
     else pkgs.haskell-nix;
-
-  projectSrc = haskell-nix.haskellLib.cleanGit {
-    name = "tezos-btc";
-    src = ./.;
-  };
 
   # haskell.nix does not support 'include' in package.yaml, we have to generate .cabal ourselves
   cabalFile = pkgs.runCommand "tzbtc.cabal" {} ''
@@ -32,20 +28,14 @@ with rec {
     ignorePackageYaml = true;
     modules =
       [{
-        packages.tzbtc = {
+        packages.tzbtc = collect-hie release {
           ghcOptions = with pkgs.lib; concatLists [
             # error on warning
             [ "-Werror" ]
 
             # disable optimizations in non-release build
             (optional (!release) "-O0")
-
-            # output *.dump-hi files (required for weeder)
-            (optionals (!release) [ "-ddump-to-file" "-ddump-hi" ])
           ];
-
-          # collect all *.dump-hi files (required for weeder)
-          postInstall = if release then "" else weeder-hacks.collect-dump-hi-files;
 
           # the code expects commit sha and date to be provided in build-time
           preBuild = if commitInfo != null then ''
@@ -55,12 +45,11 @@ with rec {
             export MORLEY_DOC_GIT_COMMIT_SHA="UNSPECIFIED"
             export MORLEY_DOC_GIT_COMMIT_DATE="UNSPECIFIED"
           '';
+          doHaddock = true;
         };
 
         # enable haddock for local package but not for dependencies
         doHaddock = false;
-        packages.tzbtc.doHaddock = true;
-
       }] ++ pkgs.lib.optional static {
         packages.tzbtc = {
           # library: no extra flags